refactor: extract comparison chart painter
